requirements.txt
weaviate-client
flask
FlagEmbedding
numpy
torch
requests
httpx
openai
together
dlai_grader
jupyter
notebook
jupyterlab
